cyber-threat/organizations/requirements.txt
py-solc-x
web3 
pandas 
scikit-learn 
numpy